+++ lib/whois/record/parser/ccwhois.ksregistry.net.rb
@@ -7,56 +7,20 @@
 #++
 
 
-require 'whois/record/parser/base'
+require 'whois/record/parser/base_shared3'
 
 
 module Whois
   class Record
     class Parser
 
+
+      # Parser for the ccwhois.ksregistry.net server.
       #
-      # = whois.adamsnames.tc
+      # @see Whois::Record::Parser::Example
+      #   The Example parser for the list of all available methods.
       #
-      # Parser for the whois.adamsnames.tc server.
-      #
-      # NOTE: This parser is just a stub and provides only a few basic methods
-      # to check for domain availability and get domain status.
-      # Please consider to contribute implementing missing methods.
-      # See WhoisNicIt parser for an explanation of all available methods
-      # and examples.
-      #
-      class WhoisAdamsnamesTc < Base
-
-        property_supported :status do
-          if available?
-            :available
-          else
-            :registered
-          end
-        end
-
-        property_supported :available? do
-          !!(content_for_scanner =~ /is not registered/)
-        end
-
-        property_supported :registered? do
-          !available?
-        end
-
-
-        property_not_supported :created_on
-
-        property_not_supported :updated_on
-
-        property_not_supported :expires_on
-
-
-        property_supported :nameservers do
-          content_for_scanner.scan(/\s+ns\s+(.+?)\s+\n/).flatten.map do |name|
-            Record::Nameserver.new(:name => name)
-          end
-        end
-
+      class CcwhoisKsregistryNet < BaseShared3
       end
 
     end